requirements_sac.txt
annotated-types==0.7.0
anyio==4.12.1
beautifulsoup4==4.14.3
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.6
curl_cffi==0.13.0
distro==1.9.0
frozendict==2.4.7
greenlet==3.3.2
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
jiter==0.13.0
multitasking==0.0.12
numpy==2.4.3
openai==2.28.0
pandas==3.0.1
peewee==4.0.2
platformdirs==4.9.4
playwright==1.58.0
protobuf==7.34.1
pycparser==3.0
pydantic==2.12.5
pydantic_core==2.41.5
pyee==13.0.1
python-dateutil==2.9.0.post0
python-dotenv==1.2.2
python-telegram-bot==22.7
pytz==2026.1.post1
requests==2.32.5
six==1.17.0
sniffio==1.3.1
soupsieve==2.8.3
tqdm==4.67.3
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.6.3
websockets==16.0
yfinance==1.2.0
